pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = ""
authors = ["Anton Petrov"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
scipy = "1.7.3"
SQLAlchemy = "^1.4.35"
ipython = "^8.2.0"
matplotlib = "^3.5.1"
pandas = "^1.4.2"
psycopg2 = "^2.9.3"
psutil = "^5.9.0"
seaborn = "^0.11.2"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"
python-dotenv = "^0.20.0"
SQLAlchemy-Utils = "^0.38.2"
pytest-factoryboy = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py39"]
exclude = '''
/(
      \.git
    | env
)/
'''

[tool.isort]
line_length = 120
skip = ["env", ".venv", "venv"]
profile = "black"

[tool.pytest]
addopts = "--cov-config=.coveragerc"
